feat: store configuration in standard folder
Without this patch, the app always tries to write its settings next to the executable.

However, if the app is installed in a non-writable path, it can't do that.

This change makes it store the configuration in the standard folder for app configuration defined by your OS. This way, there can be downstream distribution packages.